Both longtrip and longtrip2 use uint64_t now

+++ generate.h
@@ -1,5 +1,4 @@
 /*
 * Function to generate array with test hotels
 */
-void generate_hotels(int *hotels, int n);
-
+void generate_hotels(uint64_t *hotels, int n);
